+++ packages/svelte/tests/runtime-runes/samples/async-binding-update-while-focused-3/main.svelte
@@ -12,7 +12,7 @@
 </script>
 
 <button onclick={() => {
-	select.focus();
+	select?.focus();
 	resolvers.shift()?.();
 }}>shift</button>