+++ packages/svelte/tests/runtime-runes/samples/async-effect-after-await/_config.js
@@ -3,7 +3,8 @@ import { test } from '../../test';
 
 export default test({
 	async test({ assert, logs }) {
+		assert.deepEqual(logs, []);
 		await tick();
-		assert.deepEqual(logs, ['hello']);
+		assert.deepEqual(logs, ['before', 'after']);
 	}
 });